Project_2/src/requirements.txt
boto3
opencv-python
pillow
facenet-pytorch
--find-links https://download.pytorch.org/whl/cpu
torch==2.1.2